chore: refactor pages dev to use unstable_dev
