+++ .changeset/dry-hornets-jam.md
@@ -1,0 +1,5 @@
+---
+"wrangler": patch
+---
+
+chore: refactor pages dev to use unstable_dev
